[refactor] nav to seekhelp
