+++ src/screens/SeekHelp/styles.ts
@@ -2,27 +2,71 @@ import { StyleSheet } from 'react-native';
 
 export const styles = StyleSheet.create({
   container: {
-    width: '100%',
-    height: '100%',
-    alignItems: 'center',
-    justifyContent: 'center',
-    backgroundColor: '#f5f5f5',
+    display: 'flex',
     flexDirection: 'row',
-    columnGap: 20,
+    height: '100%',
   },
-  pagebutton: {
-    borderColor: '#e8e8e8',
+  leftPanel: {
+    flexDirection: 'column',
+    width: '20%',
+    backgroundColor: '#f7f9fc',
+    paddingHorizontal: 10,
+  },
+  rightPanel: {
+    display: 'flex',
+    flexDirection: 'column',
+    width: '80%',
+    height: '100%',
     backgroundColor: '#ffffff',
-    borderWidth: 2,
-    borderRadius: 20,
-    width: 500,
-    height: 200,
+    padding: 10,
+  },
+  filterButton: {
+    backgroundColor: '#f7f9fc',
+    borderRadius: 5,
+    paddingVertical: 20,
+    paddingHorizontal: 25,
+    width: '100%',
+    alignItems: 'flex-start',
     justifyContent: 'center',
-    alignItems: 'center',
+    borderWidth: 1,
+    borderColor: '#f7f9fc',
+  },
+  selectedFilterButton: {
+    backgroundColor: '#e8e8e8',
+    borderColor: '#e8e8e8',
+    borderRadius: 10,
+    marginVertical: 10,
+    marginLeft: 5,
+    paddingVertical: 10,
+    paddingHorizontal: 20,
+    width: '95%',
+    alignItems: 'flex-start',
+    borderWidth: 1,
+    justifyContent: 'center',
+    shadowColor: '#000',
+    shadowOpacity: 0.1,
+    shadowRadius: 5,
+    shadowOffset: { width: 0, height: 2 },
+  },
+  selectedButtonText: {
+    fontSize: 16,
+    color: '#444',
+    fontWeight: '600',
   },
   buttonText: {
-    fontSize: 18,
-    color: '#333',
-    fontWeight: '500',
+    fontSize: 16,
+    color: '#444',
+  },
+  selectTextContainer: {
+    marginVertical: 10,
+    marginHorizontal: 15,
+  },
+  selectText: {
+    color: '#9B9B9B',
+    fontFamily: 'Inter',
+    fontSize: 14,
+    fontWeight: '600',
+    lineHeight: 21,
+    letterSpacing: -0.154,
   },
 });
